Updated the Kalman filter pipeline.
1. previous_timestamp_ was renamed to time_us_;
2. updating of time_us_ was moved after the Kalman filter updating.
This is because the Kalman filter updating could fail, e.g. devided
by zero. In this case, time_us_ should not be updated.
3. update and updateEKF methods were renamed to KF and EKF,
respectively. The new KF/EKF methods include both the prediction
and measurement updating steps. This is because the measurement
updating may not be performed, e.g. to avoid divided by zero. In
this case, prediction should also not be performed.
4. predict method became private.